fix: Utilisation de ValidatedInput sur la période de stage
